+++ react/components/CostCenterDetails.tsx
@@ -20,6 +20,7 @@ import { AddressRules, AddressSummary } from 'vtex.address-form'
 import { costCenterMessages as messages } from './utils/messages'
 import storageFactory from '../utils/storage'
 import { setGUID } from '../utils/addresses'
+import { validatePhoneNumber } from '../modules/formValidators'
 import { useSessionResponse } from '../modules/session'
 import NewAddressModal from './NewAddressModal'
 import EditAddressModal from './EditAddressModal'
@@ -84,6 +85,7 @@ const CostCenterDetails: FunctionComponent<RouterProps> = ({
   const [permissionsState, setPermissionsState] = useState([] as string[])
   const [loadingState, setLoadingState] = useState(false)
   const [costCenterName, setCostCenterName] = useState('')
+  const [phoneNumber, setPhoneNumber] = useState('')
   const [businessDocument, setBusinessDocument] = useState('')
   const [addresses, setAddresses] = useState([] as Address[])
   const [paymentTerms, setPaymentTerms] = useState([] as PaymentTerm[])
@@ -137,34 +139,44 @@ const CostCenterDetails: FunctionComponent<RouterProps> = ({
 
   useEffect(() => {
     if (!data?.getCostCenterByIdStorefront) return
+    const {
+      organization: fetchedOrganization,
+      name: fetchedName,
+      addresses: fetchedAddresses,
+      paymentTerms: fetchedPaymentTerms = [],
+      phoneNumber: fetchedPhoneNumber = '',
+      businessDocument: fetchedBusinessDocument = '',
+    } = data.getCostCenterByIdStorefront
 
-    setCostCenterName(data.getCostCenterByIdStorefront.name)
-    handleSetAddresses(data.getCostCenterByIdStorefront.addresses)
-    setPaymentTerms(
-      data?.getCostCenterByIdStorefront?.paymentTerms?.length
-        ? data?.getCostCenterByIdStorefront?.paymentTerms
-        : []
-    )
-    if (data?.getCostCenterByIdStorefront?.businessDocument) {
-      setBusinessDocument(data.getCostCenterByIdStorefront.businessDocument)
+    setCostCenterName(fetchedName)
+    handleSetAddresses(fetchedAddresses)
+    setPaymentTerms(fetchedPaymentTerms)
+
+    if (fetchedPhoneNumber) {
+      setPhoneNumber(fetchedPhoneNumber)
+    }
+
+    if (fetchedBusinessDocument) {
+      setBusinessDocument(fetchedBusinessDocument)
     }
 
     getOrganization({
-      variables: { id: data.getCostCenterByIdStorefront.organization },
+      variables: { id: fetchedOrganization },
     })
   }, [data])
 
   useEffect(() => {
-    const termOptions = organizationData?.getOrganizationByIdStorefront
-      ?.paymentTerms?.length
-      ? organizationData.getOrganizationByIdStorefront.paymentTerms
-      : []
+    if (!organizationData?.getOrganizationByIdStorefront) return
 
-    setPaymentTermOptions(termOptions)
+    const {
+      paymentTerms: fetchedPaymentTermOptions = [],
+    } = organizationData.getOrganizationByIdStorefront
+
+    setPaymentTermOptions(fetchedPaymentTermOptions)
 
     // enable all available payment terms by default
-    if (!paymentTerms.length) {
-      setPaymentTerms(termOptions)
+    if (paymentTerms && !paymentTerms.length) {
+      setPaymentTerms(fetchedPaymentTermOptions)
     }
   }, [organizationData])
 
@@ -193,6 +205,7 @@ const CostCenterDetails: FunctionComponent<RouterProps> = ({
           return item
         }),
         paymentTerms,
+        phoneNumber,
         businessDocument,
       },
     }
@@ -460,6 +473,7 @@ const CostCenterDetails: FunctionComponent<RouterProps> = ({
                 !costCenterName ||
                 !addresses.length ||
                 (paymentTermOptions.length > 0 && paymentTerms.length === 0) ||
+                (phoneNumber && !validatePhoneNumber(phoneNumber)) ||
                 !permissionsState.includes('create-cost-center-organization')
               }
               onClick={() => handleUpdateCostCenter()}
@@ -494,21 +508,37 @@ const CostCenterDetails: FunctionComponent<RouterProps> = ({
           }
           required
         />
-      </PageBlock>
-      <PageBlock>
-        <Input
-          autocomplete="off"
-          size="large"
-          label={formatMessage(messages.businessDocument)}
-          value={businessDocument}
-          helpText={formatMessage(messages.businessDocumentHelp)}
-          onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
-            setBusinessDocument(e.target.value)
-          }}
-          readOnly={
-            !permissionsState.includes('create-cost-center-organization')
-          }
-        />
+        <div className="mt6">
+          <Input
+            autocomplete="off"
+            size="large"
+            label={formatMessage(messages.phoneNumber)}
+            value={phoneNumber}
+            helpText={formatMessage(messages.phoneNumberHelp)}
+            error={phoneNumber && !validatePhoneNumber(phoneNumber)}
+            onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
+              setPhoneNumber(e.target.value)
+            }}
+            readOnly={
+              !permissionsState.includes('create-cost-center-organization')
+            }
+          />
+        </div>
+        <div className="mt6">
+          <Input
+            autocomplete="off"
+            size="large"
+            label={formatMessage(messages.businessDocument)}
+            value={businessDocument}
+            helpText={formatMessage(messages.businessDocumentHelp)}
+            onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
+              setBusinessDocument(e.target.value)
+            }}
+            readOnly={
+              !permissionsState.includes('create-cost-center-organization')
+            }
+          />
+        </div>
       </PageBlock>
       {paymentTermOptions.length > 0 && (
         <PageBlock